Reorganize folder structure
Apply folder structure to components folder for better organization. Also moved loadPaths and handleRegionSelect functions to separate files.
