+++ src/components/infoTab/infoTab.css
