+++ src/components/quizBox/quizBox.js
