+++ src/components/statusBar/statusBar.css
